Navigation changes in inventory
Now after pressing the arrowBack button while in the navigation screen it instantly goes back to the previous screen before going to inventory.
