+++ app/src/main/java/com/github/k409/fitflow/ui/navigation/NavRoutes.kt
@@ -4,14 +4,13 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Adjust
-import androidx.compose.material.icons.outlined.Anchor
 import androidx.compose.material.icons.outlined.ErrorOutline
 import androidx.compose.material.icons.outlined.Home
+import androidx.compose.material.icons.outlined.Inventory2
 import androidx.compose.material.icons.outlined.PersonOutline
 import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material.icons.outlined.Store
 import androidx.compose.material.icons.outlined.ViewTimeline
-import androidx.compose.material.icons.outlined.Water
 import androidx.compose.material.icons.outlined.WaterDrop
 import androidx.compose.ui.graphics.vector.ImageVector
 import com.github.k409.fitflow.R
@@ -32,8 +31,7 @@ sealed class NavRoutes(
             Login,
             ProfileCreation,
             Settings,
-            Fishes,
-            Decorations,
+            Inventory,
         )
         val bottomNavBarItems = listOf(Home, Activity, Hydration, Goals, Marketplace)
     }
@@ -65,6 +63,5 @@ sealed class NavRoutes(
 
     data object Settings : NavRoutes("settings", R.string.settings, Icons.Outlined.Settings)
 
-    data object Fishes : NavRoutes("Fishes", R.string.fishes, Icons.Outlined.Water)
-    data object Decorations : NavRoutes("Decorations", R.string.decorations, Icons.Outlined.Anchor)
+    data object Inventory : NavRoutes("inventory", R.string.inventory, Icons.Outlined.Inventory2)
 }
