--- app/src/main/java/com/github/k409/fitflow/ui/screens/inventory/DecorationInventoryScreen.kt
@@ -1,106 +1,0 @@
-package com.github.k409.fitflow.ui.screens.inventory
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Anchor
-import androidx.compose.material.icons.filled.Water
-import androidx.compose.material.icons.outlined.Anchor
-import androidx.compose.material.icons.outlined.Water
-import androidx.compose.material3.Icon
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import com.github.k409.fitflow.R
-
-@Composable
-fun InventoryDecorations(navController: NavController){
-    Box(
-        modifier = Modifier.fillMaxSize()
-    ){
-        Image(
-            painter = painterResource(id = R.drawable.chest),
-            contentDescription = "Background Image",
-            modifier = Modifier.fillMaxSize(),
-        )
-    }
-
-    val ownedItems = listOf(
-        Item("Plant",
-            "I don't know what type of plant this is",
-            R.drawable.plant),
-    )
-
-    val items = listOf(
-        BottomNavigationItem(
-            title = "Fishes",
-            selectedIcon = Icons.Filled.Water,
-            unselectedIcon = Icons.Outlined.Water,
-        ),
-        BottomNavigationItem(
-            title = "Decorations",
-            selectedIcon = Icons.Filled.Anchor,
-            unselectedIcon = Icons.Outlined.Anchor,
-        )
-    )
-    var selectedItemIndex by rememberSaveable { mutableIntStateOf(0) }
-    Surface(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        Scaffold(
-            bottomBar = {
-                NavigationBar {
-                    items.forEachIndexed { index, item ->
-                        NavigationBarItem(
-                            selected = false,
-                            onClick = {
-                                selectedItemIndex = index
-                                navController.navigate(item.title)
-                            },
-                            label = {
-                                Text(text = item.title)
-                            },
-                            icon = {
-                                Icon(
-                                    imageVector = if(index == selectedItemIndex) {
-                                        item.selectedIcon
-                                    } else item.unselectedIcon,
-                                    contentDescription = item.title,
-                                )
-                            }
-                        )
-                    }
-                }
-            }
-        ) {paddingValues ->
-            Column(
-                modifier = Modifier
-                .fillMaxSize()
-                .padding(bottom = paddingValues.calculateBottomPadding())
-            ){
-                Column(
-                    modifier = Modifier
-                        .padding(horizontal = 16.dp),
-                    verticalArrangement = Arrangement.Top,
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ){List(ownedItems)}
-            }
-        }
-    }
-}
