--- app/src/main/java/com/github/k409/fitflow/ui/screens/inventory/Inventory.kt
@@ -1,9 +1,0 @@
-package com.github.k409.fitflow.ui.screens.inventory
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class BottomNavigationItem(
-    val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
-)
